fix: move CustomArtifact class from case entity and added a reference base entity for it
