+++ modelon/impact/client/entities/case.py
@@ -1,13 +1,12 @@
 from __future__ import annotations
 
 import logging
-import os
-import tempfile
 from datetime import datetime
 from typing import TYPE_CHECKING, Any, Dict, List, Optional, Text, Tuple, Union
 
 from modelon.impact.client import exceptions
 from modelon.impact.client.entities.asserts import assert_successful_operation
+from modelon.impact.client.entities.custom_artifact import CustomArtifact
 from modelon.impact.client.entities.custom_function import CustomFunction
 from modelon.impact.client.entities.external_result import ExternalResult
 from modelon.impact.client.entities.interfaces.case import CaseReference
@@ -24,7 +23,6 @@ from modelon.impact.client.operations.case import CaseOperation
 from modelon.impact.client.sal.experiment import ResultFormat
 
 if TYPE_CHECKING:
-    from modelon.impact.client.sal.experiment import ExperimentService
     from modelon.impact.client.sal.service import Service
 
 logger = logging.getLogger(__name__)
@@ -207,85 +205,6 @@ class CaseAnalysis:
     @simulation_log_level.setter
     def simulation_log_level(self, simulation_log_level: str) -> None:
         self._analysis["simulation_log_level"] = simulation_log_level
-
-
-class CustomArtifact:
-    """CustomArtifact class."""
-
-    def __init__(
-        self,
-        workspace_id: str,
-        experiment_id: str,
-        case_id: str,
-        artifact_id: str,
-        download_as: str,
-        exp_sal: ExperimentService,
-    ):
-        self._workspace_id = workspace_id
-        self._exp_id = experiment_id
-        self._case_id = case_id
-        self._artifact_id = artifact_id
-        self._download_as = download_as
-        self._exp_sal = exp_sal
-
-    @property
-    def id(self) -> str:
-        """Id of the custom artifact."""
-        return self._artifact_id
-
-    @property
-    def download_as(self) -> str:
-        """File name for the downloaded artifact."""
-        return self._download_as
-
-    def download(self, path: Optional[str] = None) -> str:
-        """Downloads a custom artifact. Returns the local path to the downloaded
-        artifact.
-
-        Args:
-            path: The local path to the directory to store the downloaded custom
-                artifact. Default: None. If no path is given, custom artifact
-                will be downloaded in a temporary directory.
-
-        Returns:
-            path: Local path to the downloaded custom artifact.
-
-        Example::
-
-            artifact_path = artifact.download()
-            artifact_path = artifact.download('/home/Downloads')
-
-        """
-        artifact, _ = self._exp_sal.case_artifact_get(
-            self._workspace_id, self._exp_id, self._case_id, self.id
-        )
-        if path is None:
-            path = os.path.join(tempfile.gettempdir(), "impact-downloads")
-        os.makedirs(path, exist_ok=True)
-        artifact_path = os.path.join(path, self.download_as)
-        with open(artifact_path, mode="wb") as f:
-            f.write(artifact)
-        return artifact_path
-
-    def get_data(self) -> Union[Text, bytes]:
-        """Returns the custom artifact stream.
-
-        Returns:
-            artifact: The artifact byte stream.
-
-        Example::
-
-            artifact = case.get_artifact("ABCD")
-            data = artifact.get_data() # may raise exception on communication error
-            with open(artifact.download_as, "wb") as f:
-                f.write(data)
-
-        """
-        result_stream, _ = self._exp_sal.case_artifact_get(
-            self._workspace_id, self._exp_id, self._case_id, self.id
-        )
-
-        return result_stream
 
 
 class CaseMeta:
